requirements.txt
fastapi
uvicorn
torch
python-multipart
pillow
torchvision